Modificação das views da Excurção

+++ Models/ExcursaoModel.cs
@@ -22,7 +22,7 @@ namespace ReserveSystem.Models
 		[Column(TypeName = "DATE")]
 		public DateTime Data_Fim { get; set;}
 		[Required,DataType(DataType.Currency), Display(Name = "Preço"), DisplayFormat(DataFormatString = "{0:C}")]
-		[Column(TypeName = "INTEGER")]
+		[Column(TypeName = "FLOAT")]
 		public float Preco {  get; set;}
 		[Required, Display(Name = "Staff ID")]
 		[Column(TypeName = "INTEGER")]
